Use AES-CTR for re-encrypting.
Change the recrypt hook to only restore the encrypted record by
re-encrypting the decrypted portion with AES-CTR.  The fixed mbuf is
then passed to the normal decrypt hook.

Allocate a separate AES-CTR session as a recrypt_sid session for
AES-GCM sessions.  When recrypting, allocate a flat buffer and encrypt
the received mbuf into the flat buffer.  Then copy back bytes from the
flat buffer back into any decrypted mbufs.

This differs a bit from what I initially proposed in that it doesn't
encrypt zero and then XOR it back.  Instead, reuse the XOR from the
encryption itself so that fixing up the M_DECRYPTED mbufs can be done
with a simple memcpy().

One possible optimization here would be to avoid allocating the
temporary char buffer for each operation by instead having each ktls
worker thread allocate a single 16k buffer on startup and pass a
pointer to that buffer down to the recrypt hook.

Make use of this refactoring to add separate counts for recrypt
operations.  In addition, this only supports AES-GCM (it seems
unlikely any NICs do Chacha20 offload?), so a new helper function is
called when allocating a receive tag to verify recrypt is supported
before allocating the tag.