Optimize for the case that there's a single decrypted region at the start.
I believe this should always be true actually, and in this case, we
can re-encrypt the mbuf chain directly without having to use a
separate output buffer.

Add some counters to see if other cases occur during testing.  If they
don't I would suggest removing support for the other cases and just
assuming this is the only case (so no need for the fixup call or
allocating a buffer, etc.  We could also in that case figure out the
offset earlier when checking the mbuf in uipc_ktls.c and pass the
decrypted length down directly to avoid a second mbuf chain walk).