cxgbei: Use the MTU to compute the data segment lengths.
By default, size the maximum receive data segment length to request
that a PDU fit within a single MSS as this permits the TOE to
avoid having to make an extra copy of received data before sending
it to the host.

Note that using smaller PDUs can add extra overhead to software
initiators and targets, so a per-adapter sysctl is provided
(dev.<nexus>.<unit>.toe.iscsi_mss_dsl) which can be used to disable
this behavior and instead use the largest PDU size supported by the
adapter (which is the previous behavior).