netinet: Simplify the signature for IPv4 protocol handlers.
Multiple IPv6 protocol handlers can be called on a single packet, so
those handlers rely on returning the next IP protocol as well as
adjusting the mbuf to trim headers for each handler.  However, IPv4
protocol handlers do not do any of this and do not need that
complexity.  Simplify IPv4 handlers to pass in the mbuf pointer and
offset by value and remove the unused return value.

Sponsored by:	Chelsio Communications