Allow IPv4 protocol handlers to accept batches of packets.
Add a parallel array of flags for IPv4 protocols that can be set via a
new argument to ipproto_register.  The only flag for now is
IPPROTO_FLAG_BATCHES.  If this flag is set, then the netisr handler
for IPv4 will coalesce batches of packets with the same IP protocol
and header length.