udp[6]_append: Don't leak inp references if a tunnel function returns false.
Tunneling functions can return false to ignore a packet and permit it
to be enqueued normally.  In that case, the reference taken on the inp
before running the tunneling function was never dropped.  Instead,
drop the reference if the function returns false, and if this was the
last reference, free the packet and return 1.