nvmecontrol: Add connect and disconnect commands.
connect accepts an address and creates an association that is passed
off to the kernel.  A discovery connect can be used which fetches the
discovery log page and associates with all of the controllers listed
in the log page.

disconnect accepts a connected fabrics controller or namespace as
an argument and disconnects from the controller.