nvmf: Support passthrough commands via ioctl.
This required adding an I/O completion callback for synchronous
commands so that they way for the I/O to complete.

Also, rather than duplicating logic for memdesc in the request layer,
change allocate_request to take the qp and allocate the capsule.
Callers can then append data directly to the capsule for the request.
This also simplifies request dispatch as it no longer has to malloc.
The cmd is now constructed similar to the userland APIs on the stack
and passed to allocate_request.