Use devctl_detach and devctl_delete for nvmecontrol disconnect.
This avoids having to do a deferred destroy_dev during detach.  I
was going to fix that anyway to be a separate ioctl on /dev/nvmf
when I realized it would just be re-inventing devctl_detach and
devctl_delete.

The forced devctl_delete is required because nvme devices are
children of root0.  Should perhaps move them to a pseudo nvmf0
bus hung off of root0.