nvmf: Close a race in nvmf_sim when disconnecting.
xpt_freeze_simq doesn't block, so we need to explicitly handle the
race of a CCB being posted to our action routine in parallel with
a disconnect.  Slap a mutex around most of nvmf_sim_io and
requeue CCBs if the sim is disconnected.

We could perhaps reduce the window of time the mutexes are held both
in nvmf_sim_io and in the equivalent nvmf_ns routine if each
nvmf_allocate_request held a reference on the associated
nvmf_host_qpair that was "released" when the request was passsed to
nvmf_submit_request.  Would have to handle freeing a request though,
so bleh.