cxgbe: Support TCP_USE_DDP on offloaded TOE connections
When this socket option is enabled, relatively large (between 64k and
16MB) buffers are allocated and used to receive data from the remote
connection.  When data is received a wrapper M_EXT mbuf is queued to
the socket's receive buffer.  In theory this should reduce the length
of the linked list of received mbufs and allow consumers to consume
receive data in larger chunks.