x86: Add a new NT_X86_CPUID regset for ptrace and core dumps
This is really intended for use in core dumps rather than ptrace (as
debuggers can just invoke cpuid directly in userland).  Right now this
note stores the contents of the various 0xd sub-leaves that describe
the layout of the XSAVE state area.  Debuggers will make use of this
to parse the contents of the NT_X86_XSTATE register set whcih can vary
by CPU.

This could also permit core dumps to use the compact XSAVE format in
the future.