nvmf: Add robustness against transport errors
During attach/detach, transport errors could result in hangs.  Add a
way to shutdown (but not fully destroy) a queue pair including
aborting active and pending commands.  When a transport error occurs
during attach or detach, shutdown the admin queue pair to ensure any
admin commands in progress are aborted with an error.  To make this
work, treat errors when enumerating namespaces as fatal that causes
attach to fail.