sys: Stop requiring nested parentheses for KASSERT
Permit passing the panic format string and arguments as additional
arguments to KASSERT itself.  This is easier to read and can be less
confusing for style checkers, etc.

Note that to preserve API compatability for now, if only two arguments
are passed to the KASSERT, the second argument must still use nested
parantheses.  Eventually we should remove this compatability.