ipmi: Remove dubious SYSUINIT that deletes devices on unload
This should only be deleting IPMI devices added via identify routines
and not IPMI devices enumerated by buses, e.g. ACPI or PCI.  Probably
ACPI or PCI devices are already detached when this runs, but the
identify routines are also smart enough to avoid adding duplicate
devices if ipmi.ko is subsequently reloaded.

It would be nice to delete unattached devices added by the identify
method of a device driver when the driver is unloaded.  However, we
do not currently have a way to track this.  Perhaps a variable in
curthread could save the current active driver around calls to
DEVICE_IDENTIFY and we could store that driver_t in any created
devices and use this to find devices during driver unload.