ext2fs: Simplify location of entries in an HTree index
Just require a fixed size for h_info_len to permit using a simpler
assignment.

When using subobject bounds on CHERI architectures, the previous
expression had the effect of shrinking the bounds on the pointer to
only cover the 'h_info' member of 'rootp', so all of the entries were
considered out of bounds.  CHERI LLVM knows to avoid narrowing the
bounds for 'h_entries' since it is a flexible array member.

Obtained from:	CheriBSD