libutil++: New library containing C++ utility classes for use in base
This initial commit adds a freebsd::stringf function that returns a
std::string constructed using a printf format string.

The current implementation is a wrapper around vasprintf that builds
and returns a std::string.  It does involve an extra memory allocation
and copy, but avoids formatting the string (and walking the va_list)
twice as solutions using vsnprintf() do.