ctld: Convert struct auth to a C++ class
Use std::string for member strings.  Instead of storing the user name
in the structure however, use an unordered_map<> with the username as
the key for the ag_auths member of auth_group.  Add a parse error if
multiple credentials specify the same user.  Previously the code
always used the first credential when verifying nad ignored additional
credentials silently.