ctld: Convert struct auth_portal to a C++ class
Use std::string for the stored copy of the initiator portal.  Note
that this string is only used in commented out debug code to dump the
configuration.

Move the logic for comparing a socket address against a single portal
into a matches class method.

Replace the TAILQ of portals in struct auth_group with a std::list<>.